pyproject.toml
[project]
name = "trip-planner-crew"
version = "0.1.0"
description = ""
authors = [{name = "Your Name", email = "you@example.com"}]
requires-python = ">=3.10.0,<3.12"
dependencies = [
    "crewai>=0.152.0",
    "unstructured==0.10.25",
    "pyowm==3.3.0",
    "tools>=0.1.9",
    "python-dotenv==1.0.0",
    "langchain[llms,openai]>=1.0.5",
    "langchain-openai>=1.0.2",
    "langchain-community>=0.4.1",
]

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

